requirements.txt
streamlit
ipykernel
bs4
requests
undetected-chromedriver
langchain
python-dotenv
langchain_community
faiss-cpu
langgraph
langchain_google_genai